[painter/gl] move #glClearBuffer() to common code
It no longer depends on the GL implementation.
