+++ internal/painter/gl/gl_goxjs.go
@@ -10,8 +10,6 @@ import (
 
 	"github.com/fyne-io/gl-js"
 	"golang.org/x/mobile/exp/f32"
-
-	"fyne.io/fyne/v2/theme"
 )
 
 const (
@@ -100,13 +98,6 @@ func (p *painter) Init() {
 	p.logError()
 
 	p.lineProgram = Program(lineProg)
-}
-
-func (p *painter) glClearBuffer() {
-	r, g, b, a := theme.BackgroundColor().RGBA()
-	p.ctx.ClearColor(float32(r)/max16bit, float32(g)/max16bit, float32(b)/max16bit, float32(a)/max16bit)
-	p.ctx.Clear(bitColorBuffer | bitDepthBuffer)
-	p.logError()
 }
 
 func (p *painter) glScissorOpen(x, y, w, h int32) {
